change: create shared contents conditions get function
