+++ src/lib/client/Shared/Helpers/Urls.ts
@@ -1,3 +1,6 @@
+import type { OrderFilters } from '$lib/client/Feature/Contents/interface';
+import type { status } from '$lib/client/Feature/Contents/Domain/ValueObjects/BookInfo/Status';
+
 export const createUrlWithParams = (url: string, params: Record<string, string>) => {
 	const searchParams = new URLSearchParams(params);
 	return `${url}?${searchParams.toString()}`;
@@ -8,4 +11,13 @@ export const getParamValue = (params: URLSearchParams, name: string) => params.g
 export const getPageCount = (params: URLSearchParams) => {
 	const pageCount = Number(getParamValue(params, 'page'));
 	return pageCount >= 0 ? pageCount : 0;
+};
+
+export const getContentsSearchConditions = (params: URLSearchParams) => {
+	const page = getPageCount(params);
+	const status = getParamValue(params, 'status') as status;
+	const query = getParamValue(params, 'query') ?? '';
+	const order = getParamValue(params, 'order') as OrderFilters;
+
+	return { page, status, query, order };
 };
