+++ src/routes/(app)/api/bookinfo/+server.ts
@@ -1,18 +1,16 @@
 import type { BookInfo } from '$lib/client/Feature/Contents/Domain/Entities/BookInfo';
 import type { BookSearch } from '$lib/client/Feature/Search/BookSearch';
+import { getContentsSearchConditions } from '$lib/client/Shared/Helpers/Urls';
 import { validatePutBookInfo } from '$lib/client/Shared/Utils/Validation';
-import collections from '$lib/server/Feature/Contents/MongoDB/MongoDBHelper';
+import { BookInfoMongoDBResource } from '$lib/server/Feature/Contents/MongoDB/BookInfoDBResource';
 import {
 	convertBookInfoToDBModel,
 	convertBookSearchToDBModel
 } from '$lib/server/Feature/Contents/MongoDB/BookInfoModel';
-import { BookInfoMongoDBResource } from '$lib/server/Feature/Contents/MongoDB/BookInfoDBResource';
+import collections from '$lib/server/Feature/Contents/MongoDB/MongoDBHelper';
 import { json } from '@sveltejs/kit';
-import type { RequestHandler } from './$types';
 import { verifyTokenAndCollections } from '../verification';
-import { getPageCount, getParamValue } from '$lib/client/Shared/Helpers/Urls';
-import type { OrderFilters } from '$lib/client/Feature/Contents/interface';
-import type { status } from '$lib/client/Feature/Contents/Domain/ValueObjects/BookInfo/Status';
+import type { RequestHandler } from './$types';
 
 export const GET: RequestHandler = async ({ url, cookies }) => {
 	const { userId, response } = await verifyTokenAndCollections(
@@ -24,7 +22,7 @@ export const GET: RequestHandler = async ({ url, cookies }) => {
 	}
 
 	const repos = new BookInfoMongoDBResource(collections?.bookInfos!, userId!);
-	const { page, status, query, order } = getSearchConditions(url.searchParams);
+	const { page, status, query, order } = getContentsSearchConditions(url.searchParams);
 
 	const bookInfos = await repos.getBookInfos(page, { status, query, order });
 	return json(bookInfos, { status: 200 });
@@ -82,13 +80,4 @@ export const DELETE: RequestHandler = async ({ request, cookies }) => {
 
 	const repos = new BookInfoMongoDBResource(collections?.bookInfos!, userId!);
 	return await repos.delete(id);
-};
-
-const getSearchConditions = (params: URLSearchParams) => {
-	const page = getPageCount(params);
-	const status = (getParamValue(params, 'status') as status) ?? '';
-	const query = getParamValue(params, 'query') ?? '';
-	const order = (getParamValue(params, 'order') as OrderFilters) ?? '';
-
-	return { page, status, query, order };
 };
